+++ src/server/server.js
@@ -86,8 +86,8 @@ export const startServer = async (config, ssl) => {
 
 	// Default middleware
 	app.use(handlerUpload(internalConfig.uploadFileSizeLimit));
-	app.use(express.json());
-	app.use(express.urlencoded({extended: true}));
+	app.use(express.json({limit: '50mb'}));
+	app.use(express.urlencoded({limit: '50mb', extended: true}));
 	app.use(cookieParser());
 	app.use(compression());