+++ tests/server.js
@@ -138,8 +138,8 @@ export const serverStart = async (config = {}) => {
 
 	// add middleware
 	app.use(handlerUpload());
-	app.use(express.json());
-	app.use(express.urlencoded({extended: true}));
+	app.use(express.json({limit: '50mb'}));
+	app.use(express.urlencoded({limit: '50mb', extended: true}));
 	app.use(cookieParser());
 	app.use(compression());